(#93) Part 3 ScreenshotsDir

+++ Source/Tests/Paths/Setup/PathsFixture.cs
@@ -4,16 +4,20 @@ public class PathsFixture : IDisposable
 {
     public static IEnumerable<object[]> GetAllDirPaths()
     {
-        yield return new object[] { CodeImp.Bloodmasters.Paths.TempDirPath };
+        yield return new object[] { CodeImp.Bloodmasters.Paths.TempDir };
         yield return new object[] { CodeImp.Bloodmasters.Paths.DownloadedResourceDir };
         yield return new object[] { CodeImp.Bloodmasters.Paths.ConfigDirPath };
+        yield return new object[] { CodeImp.Bloodmasters.Paths.ScreenshotsDir };
     }
 
     public void Dispose()
     {
-        foreach (string dir in GetAllDirPaths().SelectMany(x => x))
+        if (!string.IsNullOrEmpty(Environment.GetEnvironmentVariable("CI")))
         {
-            Directory.Delete(dir);
+            foreach (string dir in GetAllDirPaths().SelectMany(x => x))
+            {
+                Directory.Delete(dir);
+            }
         }
     }
 }
